feat(web): multi-tab doc editor — mindmap, HTML & Excel (Univer) tabs
Build out the collaborative document editor on top of the prior collab
groundwork:

- doc-tabs: multi-tab editor (BlockNote rich-text + Univer spreadsheet),
  per-tab content, autosave, and real-time collab via Hocuspocus/Yjs.
  Tab list + content synced; legacy metadata.blocknote migrates to one tab.
- Custom BlockNote blocks: /mindmap (mind-elixir) and /HTML (sandboxed
  iframe), plus a full-page HTML embed from a picked .html file.
- Office: Excel-only spreadsheet tab via Univer (lazy-loaded, sheet preset).
  Word/PowerPoint were dropped — sheet is the single office kind. Snapshots
  persist in metadata.docTabs.univer; cell text projected into content for
  search/RAG. Create a spreadsheet from the dashboard "Create New" menu.
- Editor follows app theme + locale (BlockNote + Univer dictionaries).
- Supporting fixes carried in this WIP: auth session lifetime (1d→30d) +
  remember-email, categories rename path-update int casts, CSP allows the
  collab ws: origins, doc page height fix for full-bleed editors.

Verified: tsc clean, eslint clean, next build OK; the Excel tab was
driven in a real browser (create → grid renders → type cell → persist →
reload restores the value).

+++ apps/web/lib/categories/store.ts
@@ -202,8 +202,8 @@ export async function updateCategory(
       if (admin) {
         await client.query(
           `UPDATE categories
-           SET path = $1 || substring(path from $2),
-               depth = depth + ($3 - $4),
+           SET path = $1 || substring(path from $2::int),
+               depth = depth + ($3::int - $4::int),
                updated_at = NOW()
            WHERE organization_id = $5 AND path LIKE $6 AND id != $7`,
           [path, current.path.length + 1, depth, current.depth, ctx.organizationId, `${current.path}/%`, id],
@@ -211,8 +211,8 @@ export async function updateCategory(
       } else {
         await client.query(
           `UPDATE categories
-           SET path = $1 || substring(path from $2),
-               depth = depth + ($3 - $4),
+           SET path = $1 || substring(path from $2::int),
+               depth = depth + ($3::int - $4::int),
                updated_at = NOW()
            WHERE organization_id = $5 AND user_id = $6 AND path LIKE $7 AND id != $8`,
           [path, current.path.length + 1, depth, current.depth, ctx.organizationId, ctx.userId, `${current.path}/%`, id],
